Issue #SB-25553 fix: adding api whitelist

+++ src/app/client/src/app/modules/sourcing/components/question-list/question-list.component.ts
@@ -681,7 +681,7 @@ export class QuestionListComponent implements OnInit, AfterViewInit, OnDestroy {
   handleQuestionTabChange(questionId, actionStatus?: string) {
 
     if (_.includes(this.sessionContext.questionList, questionId) && !actionStatus) { return; }
-    if (!this.checkCurrentQuestionStatus()) { return ; }
+    if (!this.checkCurrentQuestionStatus() && !this.questionCreationChild.validateCurrentQuestion()) { return ; }
     this.sessionContext.questionList = [];
     this.sessionContext.questionList.push(questionId);
     this.selectedQuestionId = questionId;
@@ -746,8 +746,7 @@ export class QuestionListComponent implements OnInit, AfterViewInit, OnDestroy {
   }
 
   public createNewQuestion(): void {
-
-    if (!this.checkCurrentQuestionStatus()) { return ; }
+    if (!this.checkCurrentQuestionStatus() || !this.questionCreationChild.validateCurrentQuestion()) { return ; }
     this.createDefaultAssessmentItem().pipe(
       map((data: any) => {
         const questionId = data.result.node_id;
@@ -1278,8 +1277,12 @@ export class QuestionListComponent implements OnInit, AfterViewInit, OnDestroy {
   }
 
   handleBack() {
-    this.generateTelemetryEndEvent('back');
-    this.programStageService.removeLastStage();
+    if(!this.questionCreationChild.validateCurrentQuestion() || !this.checkCurrentQuestionStatus()) {
+      return;
+    } else {
+      this.generateTelemetryEndEvent('back');
+      this.programStageService.removeLastStage();
+    }
   }
